+++ src/book-review/book-review.js
@@ -74,7 +74,7 @@ function BookReview({ slide, content, run, slideDone, executionId }) {
         </div>
       </div>
 
-      <ThemeStyles id={executionId} css={slide?.themeData?.cssStyles} />
+      <ThemeStyles id={executionId} css={slide?.theme?.cssStyles} />
     </>
   );
 }
@@ -86,7 +86,7 @@ BookReview.propTypes = {
       url: PropTypes.string,
       assets: PropTypes.shape({ uri: PropTypes.string }),
     }),
-    themeData: PropTypes.shape({
+    theme: PropTypes.shape({
       cssStyles: PropTypes.string,
     }),
   }).isRequired,