+++ webpack.config.js
@@ -49,7 +49,48 @@ const entry = devMode
       travel: path.resolve(__dirname, "./src/travel/travel.js"),
     };
 
-const timestamp = new Date().getTime();
+const timestamp = new Date().getTime().toString();
+
+const transformConfig = (type) => (content) => {
+  const config = JSON.parse(content.toString());
+
+  // Base build URL with trailing slash.
+  const baseUrl = (
+    type === "develop"
+      ? process.env.DEPLOYMENT_BUILD_BASE_URL_DEVELOP ??
+        process.env.DEPLOYMENT_BUILD_BASE_URL ??
+        "https://raw.githubusercontent.com/os2display/display-templates/develop/build/"
+      : process.env.DEPLOYMENT_BUILD_BASE_URL_MAIN ??
+        process.env.DEPLOYMENT_BUILD_BASE_URL ??
+        "https://raw.githubusercontent.com/os2display/display-templates/main/build/"
+  ).replace(/\/*$/, "/");
+
+  const processPath = (processablePath) => {
+    const buildPath = processablePath.replace(
+      "https://display-templates.local.itkdev.dk/build/",
+      ""
+    );
+
+    try {
+      const url = new URL(buildPath, baseUrl);
+      url.searchParams.set("ts", timestamp);
+      return url.toString();
+    } catch (error) {
+      console.error(error);
+      return processablePath;
+    }
+  };
+
+  for (const key of ["component", "admin", "schema", "assets"]) {
+    if (config.resources[key]) {
+      config.resources[key] = Array.isArray(config.resources[key])
+        ? config.resources[key].map(processPath)
+        : processPath(config.resources[key]);
+    }
+  }
+
+  return JSON.stringify(config, null, 2);
+};
 
 const plugins = devMode
   ? [
@@ -88,19 +129,7 @@ const plugins = devMode
             to: "[name]-main[ext]",
             context: path.resolve(__dirname, "src"),
             toType: "template",
-            transform(content) {
-              return content
-                .toString()
-                .replace(".json", `.json?ts=${timestamp}`)
-                .replace(".js", `.js?ts=${timestamp}`)
-                .replace(
-                  new RegExp(
-                    "https://display-templates.local.itkdev.dk/build/",
-                    "g"
-                  ),
-                  "https://raw.githubusercontent.com/os2display/display-templates/main/build/"
-                );
-            },
+            transform: transformConfig("main"),
           },
         ],
       }),
@@ -111,19 +140,7 @@ const plugins = devMode
             to: "[name]-develop[ext]",
             context: path.resolve(__dirname, "src"),
             toType: "template",
-            transform(content) {
-              return content
-                .toString()
-                .replace(".json", `.json?ts=${timestamp}`)
-                .replace(".js", `.js?ts=${timestamp}`)
-                .replace(
-                  new RegExp(
-                    "https://display-templates.local.itkdev.dk/build/",
-                    "g"
-                  ),
-                  "https://raw.githubusercontent.com/os2display/display-templates/develop/build/"
-                );
-            },
+            transform: transformConfig("develop"),
           },
         ],
       }),